requirements.txt
psycopg2-binary
pandas
scikit-learn
xgboost
matplotlib
seaborn
joblib